Monotone-Parent: 7ce650138898448e808ecf1785eedab0ad0d3e6d
Monotone-Parent: d29498142c4a9f1d780373bd5f0d80b17f6e2c33
Monotone-Revision: 41c57dcc45ba83665003200a0ea8b1cc14d09eec

Monotone-Author: ludovic@Sophos.ca
Monotone-Date: 2012-05-30T12:46:26
Monotone-Branch: ca.inverse.sogo

+++ UI/Scheduler/UIxCalView.m
@@ -118,7 +118,7 @@
 - (NSArray *) filterAppointments:(NSArray *) _apts
 {
   NSMutableArray *filtered;
-  unsigned i, count, p, pCount;
+  NSUInteger i, count, p, pCount;
   NSString *email, *partmailsString, *state, *pEmail;
   NSDictionary *info, *primaryIdentity;
   NSArray *partmails, *partstates;
@@ -370,7 +370,7 @@
 {
   NSArray        *apts;
   NSMutableArray *filtered;
-  unsigned       i, count;
+  NSUInteger     i, count;
 
   if (allDayApts)
     return allDayApts;
@@ -532,7 +532,8 @@
   NSString *uidsString, *loc, *prevMethod, *userFolderID;
   id <WOActionResults> r;
   BOOL useGroups;
-  unsigned index;
+  NSUInteger index;
+
 
   uidsString = [self queryParameterForKey: @"userUIDString"];
   uidsString = [uidsString stringByTrimmingSpaces];
